Pass 2
My submission for Pass 2 was more on understanding how to handle data retrieved from the Fetch API. Filtered data based on different prompts (such as clue value) and grouped them accordingly. Note: Pass 2 was less on the HTML and more on the JavaScript, so check the console for the data.